rename errors.py to exceptions.py
